requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
requests
qrcode
Pillow
gunicorn
python-dotenv
cryptography